tracing: setup openinference-instrumentation-openai-agents
